[Snyk] Upgrade Libs (#17)
* fix: upgrade figlet from 1.8.2 to 1.9.2

Snyk has created this PR to upgrade figlet from 1.8.2 to 1.9.2.

See this package in npm:
figlet

See this project in Snyk:
https://app.snyk.io/org/yangwang-okta/project/1bca3a75-5895-480a-a419-05c026c18604?utm_source=github&utm_medium=referral&page=upgrade-pr

* fix: upgrade axios from 1.12.0 to 1.12.2

Snyk has created this PR to upgrade axios from 1.12.0 to 1.12.2.

See this package in npm:
axios

See this project in Snyk:
https://app.snyk.io/org/yangwang-okta/project/1bca3a75-5895-480a-a419-05c026c18604?utm_source=github&utm_medium=referral&page=upgrade-pr

* fix: upgrade inquirer from 12.9.4 to 12.9.6

Snyk has created this PR to upgrade inquirer from 12.9.4 to 12.9.6.

See this package in npm:
inquirer

See this project in Snyk:
https://app.snyk.io/org/yangwang-okta/project/1bca3a75-5895-480a-a419-05c026c18604?utm_source=github&utm_medium=referral&page=upgrade-pr

* fix: upgrade puppeteer from 24.19.0 to 24.21.0

Snyk has created this PR to upgrade puppeteer from 24.19.0 to 24.21.0.

See this package in npm:
puppeteer

See this project in Snyk:
https://app.snyk.io/org/yangwang-okta/project/1bca3a75-5895-480a-a419-05c026c18604?utm_source=github&utm_medium=referral&page=upgrade-pr

---------

Co-authored-by: snyk-bot <snyk-bot@snyk.io>