Upgrade Packages 20251022 (#32)
* fix: upgrade puppeteer from 24.21.0 to 24.22.3

Snyk has created this PR to upgrade puppeteer from 24.21.0 to 24.22.3.

See this package in npm:
puppeteer

See this project in Snyk:
https://app.snyk.io/org/yangwang-okta/project/1bca3a75-5895-480a-a419-05c026c18604?utm_source=github&utm_medium=referral&page=upgrade-pr

* fix: upgrade winston from 3.17.0 to 3.18.1

Snyk has created this PR to upgrade winston from 3.17.0 to 3.18.1.

See this package in npm:
winston

See this project in Snyk:
https://app.snyk.io/org/yangwang-okta/project/1bca3a75-5895-480a-a419-05c026c18604?utm_source=github&utm_medium=referral&page=upgrade-pr

* fix: upgrade i18n from 0.15.1 to 0.15.2

Snyk has created this PR to upgrade i18n from 0.15.1 to 0.15.2.

See this package in npm:
i18n

See this project in Snyk:
https://app.snyk.io/org/yangwang-okta/project/1bca3a75-5895-480a-a419-05c026c18604?utm_source=github&utm_medium=referral&page=upgrade-pr

---------

Co-authored-by: snyk-bot <snyk-bot@snyk.io>