Upgrade libs (#54)
* fix: upgrade puppeteer from 24.23.0 to 24.25.0

Snyk has created this PR to upgrade puppeteer from 24.23.0 to 24.25.0.

See this package in npm:
puppeteer

See this project in Snyk:
https://app.snyk.io/org/yangwang-okta/project/1bca3a75-5895-480a-a419-05c026c18604?utm_source=github&utm_medium=referral&page=upgrade-pr

* fix: upgrade inquirer from 12.9.6 to 12.10.0

Snyk has created this PR to upgrade inquirer from 12.9.6 to 12.10.0.

See this package in npm:
inquirer

See this project in Snyk:
https://app.snyk.io/org/yangwang-okta/project/1bca3a75-5895-480a-a419-05c026c18604?utm_source=github&utm_medium=referral&page=upgrade-pr

* fix: upgrade semver from 7.7.2 to 7.7.3

Snyk has created this PR to upgrade semver from 7.7.2 to 7.7.3.

See this package in npm:
semver

See this project in Snyk:
https://app.snyk.io/org/yangwang-okta/project/1bca3a75-5895-480a-a419-05c026c18604?utm_source=github&utm_medium=referral&page=upgrade-pr

---------

Co-authored-by: snyk-bot <snyk-bot@snyk.io>