upgrade libs (#62)
* fix: upgrade puppeteer from 24.25.0 to 24.28.0

Snyk has created this PR to upgrade puppeteer from 24.25.0 to 24.28.0.

See this package in npm:
puppeteer

See this project in Snyk:
https://app.snyk.io/org/yangwang-okta/project/1bca3a75-5895-480a-a419-05c026c18604?utm_source=github&utm_medium=referral&page=upgrade-pr

* fix: upgrade axios from 1.12.2 to 1.13.1

Snyk has created this PR to upgrade axios from 1.12.2 to 1.13.1.

See this package in npm:
axios

See this project in Snyk:
https://app.snyk.io/org/yangwang-okta/project/1bca3a75-5895-480a-a419-05c026c18604?utm_source=github&utm_medium=referral&page=upgrade-pr

* fix: upgrade i18n from 0.15.2 to 0.15.3

Snyk has created this PR to upgrade i18n from 0.15.2 to 0.15.3.

See this package in npm:
i18n

See this project in Snyk:
https://app.snyk.io/org/yangwang-okta/project/1bca3a75-5895-480a-a419-05c026c18604?utm_source=github&utm_medium=referral&page=upgrade-pr

* bump version

---------

Co-authored-by: snyk-bot <snyk-bot@snyk.io>